[develop] Create branch develop and remove ts
