+++ npm-debug.log
@@ -1,0 +1,48 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/home/gamersclub/.nvm/versions/node/v6.17.0/bin/node',
+1 verbose cli   '/home/gamersclub/.nvm/versions/node/v6.17.0/bin/npm',
+1 verbose cli   'run',
+1 verbose cli   'dev' ]
+2 info using npm@3.10.10
+3 info using node@v6.17.0
+4 verbose run-script [ 'predev', 'dev', 'postdev' ]
+5 info lifecycle react-boilerplate@1.0.0~predev: react-boilerplate@1.0.0
+6 silly lifecycle react-boilerplate@1.0.0~predev: no script for predev, continuing
+7 info lifecycle react-boilerplate@1.0.0~dev: react-boilerplate@1.0.0
+8 verbose lifecycle react-boilerplate@1.0.0~dev: unsafe-perm in lifecycle true
+9 verbose lifecycle react-boilerplate@1.0.0~dev: PATH: /home/gamersclub/.nvm/versions/node/v6.17.0/lib/node_modules/npm/bin/node-gyp-bin:/home/gamersclub/github/react-boilerplate/node_modules/.bin:/home/gamersclub/.nvm/versions/node/v6.17.0/bin:/home/gamersclub/.npm-global/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/home/gamersclub/.rvm/bin
+10 verbose lifecycle react-boilerplate@1.0.0~dev: CWD: /home/gamersclub/github/react-boilerplate
+11 silly lifecycle react-boilerplate@1.0.0~dev: Args: [ '-c', 'webpack-dev-server --mode development' ]
+12 silly lifecycle react-boilerplate@1.0.0~dev: Returned: code: 1  signal: null
+13 info lifecycle react-boilerplate@1.0.0~dev: Failed to exec dev script
+14 verbose stack Error: react-boilerplate@1.0.0 dev: `webpack-dev-server --mode development`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (/home/gamersclub/.nvm/versions/node/v6.17.0/lib/node_modules/npm/lib/utils/lifecycle.js:255:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (/home/gamersclub/.nvm/versions/node/v6.17.0/lib/node_modules/npm/lib/utils/spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:920:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:230:5)
+15 verbose pkgid react-boilerplate@1.0.0
+16 verbose cwd /home/gamersclub/github/react-boilerplate
+17 error Linux 4.15.0-46-generic
+18 error argv "/home/gamersclub/.nvm/versions/node/v6.17.0/bin/node" "/home/gamersclub/.nvm/versions/node/v6.17.0/bin/npm" "run" "dev"
+19 error node v6.17.0
+20 error npm  v3.10.10
+21 error code ELIFECYCLE
+22 error react-boilerplate@1.0.0 dev: `webpack-dev-server --mode development`
+22 error Exit status 1
+23 error Failed at the react-boilerplate@1.0.0 dev script 'webpack-dev-server --mode development'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the react-boilerplate package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     webpack-dev-server --mode development
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs react-boilerplate
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls react-boilerplate
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
