--- src/store/ducks/roadmaps/types.ts
@@ -1,31 +1,0 @@
-/**
- * Action types
- */
-enum RoadmapsTypes {
-  LOAD_REQUEST = '@roadmaps/LOAD_REQUEST',
-  LOAD_SUCCESS = '@roadmaps/LOAD_SUCCESS',
-  LOAD_FAILURE = '@roadmaps/LOAD_FAILURE'
-}
-
-/**
- * Data types
- */
-interface Roadmap {
-  id: number,
-  name: string
-}
-
-/**
- * State types
- */
-interface RoadmapsState {
-  readonly data: Roadmap[]
-  readonly loading: boolean
-  readonly error: boolean
-}
-
-export {
-  RoadmapsTypes,
-  Roadmap,
-  RoadmapsState
-}
