--- src/store/ducks/rootReducer.ts
@@ -1,11 +1,0 @@
-import { combineReducers } from 'redux'
-
-import { roadmapsReducers } from './roadmaps/reducers'
-
-const rootReducer = combineReducers({
-  roadmapsReducers
-})
-
-export {
-  rootReducer
-}
